+++ server/extension/sql/200_quotas.sql
@@ -6,8 +6,7 @@ BEGIN
     CREATE TYPE cdb_dataservices_server.service_type AS ENUM (
       'isolines',
       'hires_geocoder',
-      'routing',
-      'observatory'
+      'routing'
     );
   END IF;
 END $$;